+++ gen/generator.go
@@ -173,17 +173,6 @@ func (g *generator) LookupConstantName(c *compile.Constant) (string, error) {
 	return name, nil
 }
 
-// servicePackage imports the service package for the given service and returns
-// the name that should be used to refer to that package.
-func (g *generator) servicePackage(s *compile.ServiceSpec) (string, error) {
-	pkg, err := g.thriftImporter.ServicePackage(s.ThriftFile(), s.Name)
-	if err != nil {
-		return "", err
-	}
-
-	return g.Import(pkg), nil
-}
-
 // TextTemplate renders the given template with the given template context.
 func (g *generator) TextTemplate(s string, data interface{}, opts ...TemplateOption) (string, error) {
 	templateFuncs := template.FuncMap{
@@ -195,7 +184,6 @@ func (g *generator) TextTemplate(s string, data interface{}, opts ...TemplateOpt
 		"isStructType":     isStructType,
 		"newNamespace":     g.Namespace.Child,
 		"newVar":           g.Namespace.Child().NewName,
-		"servicePackage":   g.servicePackage,
 		"typeName":         curryGenerator(typeName, g),
 		"typeReference":    curryGenerator(typeReference, g),
 		"typeReferencePtr": curryGenerator(typeReferencePtr, g),